SNOW-350686: Introduce Streaming Ingest Channel Cache (#10)
Channel Cache is basically an in-memory map where the key is a fully qualified table name and value is a set of channels that belongs to this table, so that we could combine the channel data together into one chunk during the flush. Each client will have one Channel Cache and it will contain all the active channels for that client. When users open a new channel, it will be added to the cache and if the channel exists already, the old one will be invalided and removed from the cache. When users explicitly close a channel or the channel somehow gets invalided by the server response, the channel will be removed from the cache as well. Concurrent access will be supported in the cache.