SNOW-350686: Introduce Streaming Ingest Flush Service (#13)
Flush Service contains three components (Build, Upload and Register), and together they’re responsible for flushing data from the client to Snowflake tables, and flushing could be triggered by either background periodical flush or when the number of data exceeds a certain size or users explicitly call flush on channel/client. All the work inside the flush service will happen asynchronously. And depending on the number of threads in the thread pool and the total number of data to be flushed, we will try to distribute the work evenly in order to achieve higher throughput.