SNOW-365158: Introduce InsertResponse for InsertRow API  (#37)
Introduce an InsertValidationResponse class which may contain a list of InsertError, and for each InsertError, it will include the erroneous row content and the corresponding exception. Currently, the behavior is similar to ON_ERROR=CONTINUE (skip erroneous rows and continue inserting good rows). The long term plan is to make it consistent with the ON_ERROR options available in COPY (of course, based on customer feedback), and we may also support an error stage (similar to dead-letter queue) for uploading the erroneous rows.