SNOW-382458: fix memory leak during failure cases (#52)
Spent a few days on investigating this memory leak and finally root cause it, the reason is due to async processing of the flush process + usage of futures, so if there is a timeout (or random) exception before we finishing building the blob, the vectors in the arrow buffer are not released before closing the allocator, which ends up with a memory leak issue. So the fix is basically to always release the vectors before invalidate the channels (and close the allocator).