remove public flush API (#60)
We want to remove the public flush API, in order to:
1. Simplify user experience
2. Simplify internal implementation

Of course this will add limitations for powerful users, but we would want this at least for PrPr and will gather feedbacks. Add this as a separate PR so we could easily revert it if needed in the future.