SNOW-259103: Streaming Ingest Chunk level data encryption in blob (#54)
Data Encryption
As described in https://docs.google.com/document/d/1TAcg6hqcfEKtcFV7plGSVzw9xbgNt5TsUCRkIJJ_o54/edit, the server will pass a BTMK to the client as part of the channels/open response. In the client, this BTMK will be used to derive the CBTMK (client BTMK) by using a dynamic diversifier (full blob path). Each chunk of data will be compressed first, then encrypted using the CBTMK, with AES CTR mode and zero initialization vector.

Metadata Encryption
It will be handled in the separate PR in the future, for now, we don't include any metadata in the blob, tracked by SNOW-389883