NO-SNOW: fix month generated by Calendar (#84)
Convert the month to start from 1, this is a really interesting problem

Field number for get and set indicating the month. This is a calendar-specific value. The first month of the year in the Gregorian and Julian calendars is JANUARY which is 0; the last depends on the number of months in a year.