SNOW-566797: Support SDK with proxy server  (#154)
Looks like we're not setting proxy properties as part of the uploadWithoutConnection call to upload files, this PR adds the support. Will need to see how we would setup the proxy server as part of the GH action if we want to add a real IT tests in the future.