NO-SNOW: increase retry count due to server side Commit Service not making progress  (#172)
As part of the channel.close API, we will keep checking whether the server side is making progress (committing offset), and throw an error with certain timeout if no progress. The current timeout is set to 5s, and according to our recent resilience testing, somehow the Commit Service does not able to make any progress. While the team is investigating the root cause, increase the retry timeout to 10s (retry interval * retry count) instead.