SNOW-644956: fix string data type  in ARRAY (#211)
When an array (or list) of string is passed into the Ingest SDK, the string doesn’t have double quotes around it when the array is dumped into a string, leading to a failed query later on.

Desired behavior:

[“abc”] → “[“abc”]“

[[“abc”], “de”] → “[[“abc”],”de”]“

Current behavior:

[“abc”] → “[abc]“

[[“abc”], “de”] → “[[abc],de]“

The reason is that the toString method is called to make element of the array into string, and the method will not add quotes around the string as we intend. There could also be issues with other data types whose toString method is not implemented to match our interface.