Remove the dependency on JAXB-API
It was only being used for parsing and printing md5 hex strings. JAXB is notorious for having
challenges with run-time bindings and relocations (as indicated in the StackOverflow post)
linked in the now-removed dependency.

Apache Commons Codec is already present, so we can switch to to using `o.a.c.codec.binary.Hex`
instead.