Add maven-dependency-plugin
Adding maven-dependency plugin surfaces these dependency issues,
also fixed by this PR:

- Declare used undeclared dependencies
  - com.google.code.findbugs:jsr305:3.0.2
  - org.apache.parquet:parquet-common:1.12.3
  - com.fasterxml.jackson.core:jackson-core:2.14.0
  - io.netty:netty-common:4.1.82.Final
  - com.fasterxml.jackson.core:jackson-annotations:2.14.0
  - commons-codec:commons-codec:1.15
  - org.apache.arrow:arrow-memory-core:10.0.0
  - org.apache.parquet:parquet-column:1.12.3
  - org.apache.commons:commons-lang3:3.12.0
  - com.google.guava:guava:27.0-jre
- Remove unused declared dependency
  - org.apache.hadoop:hadoop-mapreduce-client-core:3.3.4:compile

A dependency is considered used when it is imported by source java files present
in this project, and unused when not. It is good hygiene to declare a dependency
when used to avoid relying on transitive dependencies.

Related Java Library Best Practice: https://jlbp.dev/JLBP-22