SNOW-735517 Fix nullCount/rowCount mismatch (#357)
In `addRow`, the SDK iterates over columns of the row one-by-one, it
validates, calculates statistics and puts the resulting value in the
row buffer. The problem is when an exception is thrown (e.g. validation
fails), it (correctly) doesn’t flush the row, but it forgets to “undo”
the statistics calculated for the previous, already processed, columns
in the row.

This PR forks for each column during row processing and only updates
the resulting statistics when all column are processed successfully.