Fix shading class relocation: net.snowflake.ingest.shade
- Relocate all dependency classes under net.snowflake.ingest.shade

As of release 1.1.3 the shaded jar contains a large number of unrelocated
classes. This is problematic for library consumers that include any of the same
dependencies in their projects that snowflake-ingest-sdk uses in addition to
snowflake-ingest-sdk. This can result in duplicate classes in a project and
confusing errors for users depending on dependency ordering of their project.
When duplicate classes are present on the classpath the first class found wins.

This means that for consumers of snowflake-ingest-sdk if it's a dependency
listed high in a pom.xml's dependency section, its duplicate classes take
precedence, overriding any other dependency definitions in their project.

Here's a short example of the problem. If I have a project using the ingest SDK
with a pom.xml as follows:

```
      <dependency>
        <groupId>net.snowflake</groupId>
        <artifactId>snowflake-ingest-sdk</artifactId>
        <version>1.1.3</version>
      </dependency>
      <dependency>
        <groupId>com.google.protobuf</groupId>
        <artifactId>protobuf-java</artifactId>
        <version>3.22.3</version>
      </dependency>
```

snowflake-ingest-sdk includes class files from
com.google.protobuf:protobuf-java:2.5.0 which is pulled in transitively from
hadoop-common. For any class files that are duplicates of protobuf-java 3.22.3
the class files of 2.5.0 will take precedence at runtime due to
snowflake-ingest-sdk being first in the pom.xml and therefore first on the
classpath.

Conversely if snowflake-ingest-sdk is last then protobuf-java 3.22.3 will be
active which may cause runtime issues with snowflake-ingest-sdk if it depends on
APIs only present in 2.5.0.

Fixes #356