SNOW-855191 Race condition causing NPE
There is a race condition, which leads to a NPE when one thread invalidates the channel,
while another thread is inside of insertRows(). The problem is that channel invalidation
clears fieldIndex, which may still be used in threads concurrently inserting rows.

There is no need to clear the fieldIndex during invalidation, it is enough to mark
the channel as invalid and it won't be flushed. When channel is reopened, setupSchema()
will clear fieldIndex and reinitialize the schema.