v2.0.2 release  (#558)
Release v2.0.2, this will be used as our first GA release! This release is on branch https://github.com/snowflakedb/snowflake-ingest-java/tree/tzhang-si-ga since we only want to include selected changes on top of v2.0.1 to make sure it's a stable release, anything feature related will be released in v2.0.3 in the future.