SNOW-914666 Adds MAX_CLIENT_LAG configuration option (#586)
SNOW-914666 Adds MAX_CLIENT_LAG configuration option

We want to expose a knob that gives users the ability to control when
data is ingested. This has a material difference on the size of blobs
generated and can result in fewer smaller-sized blobs which in turn
affects query performance. The trade-off is higher ingest latencies. We
have decided to expose this in the form of an optional `MAX_CLIENT_LAG`
option that accepts inputs as the following:

- `number second` (ex: `1 second`)
- `number seconds` (ex: `2 seconds`)
- `number minute` (ex: `1 minute`)
- `number minutes` (ex: `2 minutes`)

By default we use 1 second as the maximum client lag which is the
current behavior of the SDK.

Note that this dictates when a flush is triggered to cloud storage.
Depending on your connection to cloud storage and cloud storage tail
latencies a blob persist may take longer than expected. Therefore, it is
helpful to think of this parameter as a target, rather than an absolute
number.

If you have manually overridden any flush-related parameters in the past
this can be disabled by setting `MAX_CLIENT_LAG_ENABLED` to `false`
in your configuration file.

@test Adds tests to `ParameterProviderTest`