SNOW-919423 Surfaces Table Schema for a Channel
We've had requests to surface the table schema for a channel so that it
is possible to reason about what columns need to be supplied to a given
call of `insertRow` or `insertRows`. This surfaces a map of Column
Name to Column Properties that are normally surfaced in the output of
`SHOW COLUMNS`.

@test adds test to `SnowflakeStreamingIngestChannelTest.java`