Fix quoted column name for Iceberg primitive type (#891)
For non-structured data type column, server sides perform EP info check which requires the column display name to be exactly the same as the quoted column name. Fix a bug on this and extend existing column name test to run on iceberg as in preprod as well. Disable the Iceberg test now the pass the merge gate.